Removing all crypto usages
