+++ hyperledger/chaincode/go/computeTally.go
@@ -3,9 +3,7 @@ package main
 import (
 	"encoding/json"
 	"fmt"
-	"math/big"
 
-	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/hyperledger/fabric/core/chaincode/shim"
 	sc "github.com/hyperledger/fabric/protos/peer"
 )
@@ -16,16 +14,14 @@ func (s *SmartContract) computeTally(stub shim.ChaincodeStubInterface) sc.Respon
 		return shim.Error("Wrong state")
 	}
 
-	curve := crypto.S256()
-
 	var totalRegistered int
 	GetState(stub, "totalregistered", &totalRegistered)
 
 	var voters map[string]Voter
 	GetState(stub, "voters", &voters)
 
-	// is a point on the curve, I guess
-	tempResult := []*big.Int{big.NewInt(0), big.NewInt(0)}
+	// Initialize all results with 0
+	tempResult := []int{0}
 
 	// Sum all votes
 	for voterAddress, voter := range voters {
@@ -39,37 +35,19 @@ func (s *SmartContract) computeTally(stub shim.ChaincodeStubInterface) sc.Respon
 			return shim.Error("Voter " + voterAddress + " has not voted")
 		}
 
-		vote := voter.vote
+		vote := voter.Vote
 		logger.Info(vote)
 
-		tempResult[0], tempResult[1] = curve.Add(tempResult[0], tempResult[1], vote[0], vote[1])
+		tempResult[vote]++
 
-		if tempResult[0] == big.NewInt(0) {
-			finalTally := []int{0, totalRegistered}
-			finalTallyBytes, _ := json.Marshal(finalTally)
-			return shim.Success(finalTallyBytes)
-		} else {
-			tempG := []*big.Int{curve.Params().Gx, curve.Params().Gy}
-			for i := 1; i <= totalRegistered; i++ {
-				if tempResult[0] == tempG[0] {
-					finalTally := []int{i, totalRegistered}
-					finalTallyBytes, _ := json.Marshal(finalTally)
-					return shim.Success(finalTallyBytes)
-				}
-				tempG[0], tempG[1] = curve.Add(tempG[0], tempG[1], curve.Params().Gx, curve.Params().Gy)
-			}
-		}
-
-		// Something bad happened, we should never get here
-		finalTally := []int{0, 0}
-		finalTallyBytes, _ := json.Marshal(finalTally)
-		return shim.Success(finalTallyBytes)
 	}
+
+	finalTally := Result{totalRegistered, tempResult}
 
 	// All votes have been accounted for...
 	// Get tally and change state to 'Finished'
 	s.transitionToState(stub, FINISHED)
 
-	return shim.Error("Not implemented yet")
-
+	finalTallyBytes, _ := json.Marshal(finalTally)
+	return shim.Success(finalTallyBytes)
 }
