+++ hyperledger/chaincode/go/loadDependencies.sh
@@ -2,4 +2,3 @@
 
 # For testing
 go get -v github.com/stretchr/testify
-go get -v github.com/ethereum/go-ethereum
