+++ hyperledger/chaincode/go/register.go
@@ -1,10 +1,6 @@
 package main
 
 import (
-	"crypto/ecdsa"
-	"encoding/json"
-	"math/big"
-
 	"github.com/hyperledger/fabric/core/chaincode/shim"
 	sc "github.com/hyperledger/fabric/protos/peer"
 )
@@ -14,29 +10,11 @@ func (s *SmartContract) register(stub shim.ChaincodeStubInterface, args []string
 		return shim.Error("Wrong state")
 	}
 
-	if len(args) != 3 {
-		return shim.Error("Wrong number of arguments, expected 3")
+	if len(args) != 1 {
+		return shim.Error("Incorrect number of arguments, expecting userId")
 	}
 
-	// TODO: what do these names mean?
-	// TODO: error handling
-
-	userIDBytes, _ := stub.GetCreator()
-	userID := string(userIDBytes)
-
-	// Public key of voter: xG
-	var xG []*big.Int
-	json.Unmarshal([]byte(args[0]), &xG)
-
-	publicKey := new(ecdsa.PublicKey)
-	publicKey.X = xG[0]
-	publicKey.Y = xG[1]
-
-	var vG []*big.Int
-	json.Unmarshal([]byte(args[1]), &vG)
-
-	var r big.Int
-	json.Unmarshal([]byte(args[2]), &r)
+	userID := args[0]
 
 	var eligible map[string]bool
 	GetState(stub, "eligible", &eligible)
@@ -47,20 +25,21 @@ func (s *SmartContract) register(stub shim.ChaincodeStubInterface, args []string
 	isEligible := eligible[userID]
 	isRegistered := registered[userID]
 
-	var emptyVote []*big.Int
-	emptyReconstructedKey := new(ecdsa.PublicKey)
-	voter := Voter{userID, publicKey, emptyReconstructedKey, emptyVote}
+	voter := Voter{userID, -1}
 
-	if isEligible && !isRegistered && s.verifyZKP(voter, &r, vG) {
+	if isEligible && !isRegistered {
 		registered[userID] = true
 		PutState(stub, "registered", registered)
-
-		// voter := Voter{userID, xG, {}, {}}
 
 		var totalRegistered int
 		GetState(stub, "totalRegistered", &totalRegistered)
 		totalRegistered = totalRegistered + 1
 		PutState(stub, "totalRegistered", totalRegistered)
+
+		var voters []Voter
+		GetState(stub, "voters", &voters)
+		voters = append(voters, voter)
+		PutState(stub, "voters", voters)
 	}
 
 	return shim.Success(nil)
