+++ hyperledger/chaincode/go/state.go
@@ -30,8 +30,6 @@ func (state StateEnum) String() string {
 func (s *SmartContract) inState(stub shim.ChaincodeStubInterface, expectedState StateEnum) bool {
 	var state StateEnum
 	GetState(stub, "state", &state)
-
-	logger.Info("State is " + state.String())
 	return expectedState == state
 }
