+++ hyperledger/chaincode/go/voting_test.go
@@ -63,60 +63,6 @@ func Test_InvalidFunctionName(t *testing.T) {
 	}
 }
 
-func Test_Question(t *testing.T) {
-	scc := new(SmartContract)
-	stub := shim.NewMockStub("test_question", scc)
-
-	question := "What is the question?"
-	stub.MockTransactionStart("t123")
-	PutState(stub, "question", question)
-	stub.MockTransactionEnd("t123")
-
-	checkQuery(t, stub, "question", question)
-}
-
-func Test_ComputeTally(t *testing.T) {
-
-	scc := new(SmartContract)
-	stub := shim.NewMockStub("test_computeTally", scc)
-
-	stub.MockTransactionStart("t123")
-	PutState(stub, "state", VOTE)
-	stub.MockTransactionEnd("t123")
-
-	checkInvoke(t, stub, [][]byte{[]byte("computeTally")})
-}
-
-func Test_Register(t *testing.T) {
-
-	scc := new(SmartContract)
-	stub := shim.NewMockStub("test_register", scc)
-
-	stub.MockTransactionStart("t123")
-	PutState(stub, "state", SIGNUP)
-	eligible := map[string]bool{"userId": true}
-	PutState(stub, "eligible", eligible)
-	registered := map[string]bool{"userId": false}
-	PutState(stub, "registered", registered)
-	stub.MockTransactionEnd("t124")
-
-	checkInvoke(t, stub, [][]byte{
-		[]byte("register"),
-		[]byte("[1, 2]"),
-		[]byte("[1, 2, 3]"),
-		[]byte("1"),
-	})
-
-	var totalRegistered int
-	GetState(stub, "totalRegistered", &totalRegistered)
-	assert.Equal(t, 1, totalRegistered)
-
-	GetState(stub, "registered", &registered)
-	assert.True(t, registered["userId"])
-
-	// TODO: verify that voter was stored
-}
-
 func Test_SetEligible(t *testing.T) {
 
 	scc := new(SmartContract)
@@ -139,59 +85,4 @@ func Test_SetEligible(t *testing.T) {
 	assert.True(t, eligible["a"])
 	assert.True(t, eligible["b"])
 	assert.True(t, eligible["c"])
-}
-
-func Test_InitVote(t *testing.T) {
-
-	scc := new(SmartContract)
-	stub := shim.NewMockStub("test_beginSignup", scc)
-
-	checkInvoke(t, stub, [][]byte{[]byte("initVote")})
-
-	var state StateEnum
-	GetState(stub, "state", &state)
-	assert.Equal(t, SETUP, state)
-}
-
-func Test_BeginSignup(t *testing.T) {
-
-	scc := new(SmartContract)
-	stub := shim.NewMockStub("test_beginSignup", scc)
-
-	stub.MockTransactionStart("t123")
-	PutState(stub, "state", SETUP)
-	stub.MockTransactionEnd("t123")
-
-	checkInvoke(t, stub, [][]byte{[]byte("beginSignUp"), []byte("This is a question!")})
-
-	var state StateEnum
-	GetState(stub, "state", &state)
-	assert.Equal(t, SIGNUP, state)
-
-	var question string
-	GetState(stub, "question", &question)
-	assert.Equal(t, "This is a question!", question)
-}
-
-func Test_SubmitVote(t *testing.T) {
-
-	scc := new(SmartContract)
-	stub := shim.NewMockStub("test_submitVote", scc)
-
-	stub.MockTransactionStart("t123")
-	PutState(stub, "state", VOTE)
-	stub.MockTransactionEnd("t123")
-
-	userId := "1"
-	stub.MockTransactionStart("t124")
-	registered := map[string]bool{userId: true}
-	PutState(stub, "registered", registered)
-	stub.MockTransactionEnd("t124")
-
-	stub.MockTransactionStart("t125")
-	votecast := map[string]bool{userId: false}
-	PutState(stub, "votecast", votecast)
-	stub.MockTransactionEnd("t125")
-
-	checkInvoke(t, stub, [][]byte{[]byte("submitVote"), []byte(userId)})
 }
