+++ hyperledger/start.sh
@@ -36,9 +36,6 @@ docker exec -e "CORE_PEER_LOCALMSPID=Org1MSP" -e "CORE_PEER_MSPCONFIGPATH=/etc/h
 # Now launch the CLI container in order to install, instantiate chaincode
 docker-compose -f ./docker-compose.yml up -d cli
 
-# Install go dependencies
-docker exec cli /opt/gopath/src/hpi.de/go/loadDependencies.sh
-
 printf "\n############### Installing chaincode\n"
 docker exec cli peer chaincode install -n vote -v 1.0 -p "$CC_SRC_PATH" -l "$LANGUAGE"
